+++ src/utils.js
@@ -1,19 +1,23 @@
+var fs = require('fs');
+var path = require('path');
+var vscode = require('vscode');
 let exec = require('child_process').exec;
+var npm = require('npm');
 
 // checks if there exists a valid installation of NodeJs on this machine
 exports.isNodeInstalled = function isNodeInstalled() {
-  var cmdString = "node -v";
-  return new Promise(function (resolve, reject) {
-    exec(cmdString, (error, stdout) => {
-      if (error) {
-        return reject(error);
-      }
-      if (stdout.startsWith('v')) {
-        return resolve(true);
-      }
-      return resolve(false);
+    var cmdString = "node -v";
+    return new Promise(function (resolve, reject) {
+        exec(cmdString, (error, stdout) => {
+            if (error) {
+                return reject(error);
+            }
+            if (stdout.startsWith('v')) {
+                return resolve(true);
+            }
+            return resolve(false);
+        });
     });
-  });
 };
 
 // lists all globally installed npm packages.
@@ -70,7 +74,8 @@ exports.npmInstall = function npmInstall(packages, opts) {
     var cmdString = "npm install " + packages.join(" ") + " "
         + (opts.global ? " -g" : "")
         + (opts.save ? " --save" : "")
-        + (opts.saveDev ? " --saveDev" : "");
+        + (opts.saveDev ? " --saveDev" : "")
+        + (opts.prefix ? " --prefix " + opts.prefix : "");
 
     return new Promise(function (resolve, reject) {
         exec(cmdString, { cwd: opts.cwd ? opts.cwd : "/" }, (error) => {
@@ -81,4 +86,53 @@ exports.npmInstall = function npmInstall(packages, opts) {
             }
         });
     });
+};
+
+exports.getPackageJsonPath = function getPackageJsonPath() {
+    var dirAboveRoot = path.join(vscode.workspace.rootPath, '..');
+    var srcPath = getSourceLocation();
+    if (!srcPath) {
+        return;
+    }
+
+    var packageJsonPath;
+    while (srcPath !== dirAboveRoot) {
+        packageJsonPath = path.join(srcPath, 'package.json');
+        if (fs.existsSync(packageJsonPath)) {
+            return packageJsonPath;
+        }
+        else {
+            srcPath = path.join(srcPath, '..');
+        }
+    }
+};
+
+function getSourceLocation() {
+    var files = vscode.workspace.textDocuments.filter(item => item.isUntitled === false);
+    if (files) {
+        var sourceFile = files[0].fileName;
+        return sourceFile.slice(0, sourceFile.lastIndexOf('\\') + 1);
+    }
+}
+
+exports.install = function install(pkgName, options) {
+    var promise = new Promise(function (resolve, reject) {
+        if (!options) {
+            reject("options.prefix is required.");
+        }
+
+        npm.load(options, function (err) {
+            if (err) {
+                return reject(err);
+            }
+            npm.commands.install([pkgName + '@latest'], function (err, info) {
+                if (err) {
+                    reject(err);
+                }
+                resolve(info);
+            });
+        });
+    });
+
+    return promise;
 };